+++ src/Schedule.h
@@ -291,10 +291,11 @@ public:
     std::vector<Bound> &estimates();
     // @}
 
-    /** Mark calls of a function by 'f' to be replaced with its wrapper
-     * during the lowering stage. If the string 'f' is empty, it means replace
-     * all calls to the function by all other functions (excluding itself) in
-     * the pipeline with the wrapper. See \ref Func::in for more details. */
+    /** Mark calls of a function by 'f' to be replaced with its identity
+     * wrapper or clone during the lowering stage. If the string 'f' is empty,
+     * it means replace all calls to the function by all other functions
+     * (excluding itself) in the pipeline with the global identity wrapper.
+     * See \ref Func::in and \ref Func::clone for more details. */
     // @{
     const std::map<std::string, Internal::FunctionPtr> &wrappers() const;
     std::map<std::string, Internal::FunctionPtr> &wrappers();
