+++ test/correctness/func_clone.cpp
