+++ test/correctness/func_wrapper.cpp
@@ -10,19 +10,19 @@ using std::string;
 using namespace Halide;
 using namespace Halide::Internal;
 
-int calling_in_in_wrapper_no_op_test() {
+int calling_wrapper_no_op_test() {
     Var x("x"), y("y");
 
     {
-        ImageParam img(Int(32), 2);
-        Func f("f");
-        f(x, y) = img(x, y);
+        Func f("f"), g("g");
+        f(x, y) = x + y;
+        g(x, y) = f(x, y);
 
-        // Calling wrap on the same ImageParam for the same Func multiple times should
+        // Calling wrap on the same Func for the same Func multiple times should
         // return the same wrapper
-        Func wrapper = img.in(f);
+        Func wrapper = f.in(g);
         for (int i = 0; i < 5; ++i) {
-            Func temp = img.in(f);
+            Func temp = f.in(g);
             if (wrapper.name() != temp.name()) {
                 std::cerr << "Expect " << wrapper.name() << "; got " << temp.name() << " instead\n";
                 return -1;
@@ -31,13 +31,13 @@ int calling_in_in_wrapper_no_op_test() {
     }
 
     {
-        ImageParam img(Int(32), 2);
-        Func f("f");
-        f(x, y) = img(x, y);
+        Func f("f"), g("g");
+        f(x, y) = x + y;
+        g(x, y) = f(x, y);
 
         // Should return the same global wrapper
-        Func wrapper1 = img.in();
-        Func wrapper2 = img.in();
+        Func wrapper1 = f.in();
+        Func wrapper2 = f.in();
         if (wrapper1.name() != wrapper2.name()) {
             std::cerr << "Expect " << wrapper1.name() << "; got " << wrapper2.name() << " instead\n";
             return -1;
@@ -45,15 +45,15 @@ int calling_in_in_wrapper_no_op_test() {
     }
 
     {
-        ImageParam img(Int(32), 2);
-        Func e("e"), f("f"), g("g"), h("h");
-        e(x, y) = img(x, y);
-        f(x, y) = img(x, y);
-        g(x, y) = img(x, y);
-        h(x, y) = img(x, y);
+        Func d("d"), e("e"), f("f"), g("g"), h("h");
+        d(x, y) = x + y;
+        e(x, y) = d(x, y);
+        f(x, y) = d(x, y);
+        g(x, y) = d(x, y);
+        h(x, y) = d(x, y);
 
-        Func wrapper1 = img.in({e, f, g});
-        Func wrapper2 = img.in({g, f, e});
+        Func wrapper1 = d.in({e, f, g});
+        Func wrapper2 = d.in({g, f, e});
         if (wrapper1.name() != wrapper2.name()) {
             std::cerr << "Expect " << wrapper1.name() << "; got " << wrapper2.name() << " instead\n";
             return -1;
@@ -63,31 +63,26 @@ int calling_in_in_wrapper_no_op_test() {
     return 0;
 }
 
-int func_in_in_wrapper_test() {
-    Func source("source"), g("g");
+int func_wrapper_test() {
+    Func f("f"), g("g");
     Var x("x"), y("y");
 
-    source(x) = x;
-    ImageParam img(Int(32), 1, "img");
-    Buffer<int> buf = source.realize(200);
-    img.set(buf);
+    f(x) = x;
+    g(x, y) = f(x);
 
-    g(x, y) = img(x);
-
-    Func wrapper = img.in(g).compute_root();
-    Func img_f = img;
-    img_f.compute_root();
+    Func wrapper = f.in(g).compute_root();
+    f.compute_root();
 
     // Check the call graphs.
-    // Expect 'g' to call 'wrapper', 'wrapper' to call 'img_f', 'img_f' to call 'img'
-    Module m = g.compile_to_module({g.infer_arguments()});
+    // Expect 'g' to call 'wrapper', 'wrapper' to call 'f', 'f' to call nothing
+    Module m = g.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
         {g.name(), {wrapper.name()}},
-        {wrapper.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {wrapper.name(), {f.name()}},
+        {f.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -101,37 +96,32 @@ int func_in_in_wrapper_test() {
     return 0;
 }
 
-int multiple_funcs_sharing_in_wrapper_test() {
-    Func source("source"), g1("g1"), g2("g2"), g3("g3");
+int multiple_funcs_sharing_wrapper_test() {
+    Func f("f"), g1("g1"), g2("g2"), g3("g3");
     Var x("x"), y("y");
 
-    source(x) = x;
-    ImageParam img(Int(32), 1, "img");
-    Buffer<int> buf = source.realize(200);
-    img.set(buf);
+    f(x) = x;
+    g1(x, y) = f(x);
+    g2(x, y) = f(x);
+    g3(x, y) = f(x);
 
-    g1(x, y) = img(x);
-    g2(x, y) = img(x);
-    g3(x, y) = img(x);
-
-    Func im_wrapper = img.in({g1, g2}).compute_root();
-    Func img_f = img;
-    img_f.compute_root();
+    f.compute_root();
+    Func f_wrapper = f.in({g1, g2}).compute_root();
 
     // Check the call graphs.
-    // Expect 'g1' and 'g2' to call 'im_wrapper', 'g3' to call 'img_f',
-    // im_wrapper' to call 'img_f', 'img_f' to call 'img'
+    // Expect 'g1' and 'g2' to call 'f_wrapper', 'g3' to call 'f',
+    // f_wrapper' to call 'f', 'f' to call nothing
     Pipeline p({g1, g2, g3});
-    Module m = p.compile_to_module({p.infer_arguments()}, "");
+    Module m = p.compile_to_module({}, "");
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {g1.name(), {im_wrapper.name()}},
-        {g2.name(), {im_wrapper.name()}},
-        {g3.name(), {img_f.name()}},
-        {im_wrapper.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {g1.name(), {f_wrapper.name()}},
+        {g2.name(), {f_wrapper.name()}},
+        {g3.name(), {f.name()}},
+        {f_wrapper.name(), {f.name()}},
+        {f.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -154,38 +144,33 @@ int multiple_funcs_sharing_in_wrapper_test() {
     return 0;
 }
 
-int global_in_in_wrapper_test() {
-    Func source("source"), g("g"), h("h"), i("i");
+int global_wrapper_test() {
+    Func f("f"), g("g"), h("h"), i("i");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
-
-    g(x, y) = img(x, y);
-    h(x, y) = g(x, y) + img(x, y);
+    f(x, y) = x + y;
+    g(x, y) = f(x, y);
+    h(x, y) = g(x, y) + f(x, y);
 
     Var xi("xi"), yi("yi"), t("t");
-    Func wrapper = img.in();
-    Func img_f = img;
-    img_f.compute_root();
+    Func wrapper = f.in();
+    f.compute_root();
     h.compute_root().tile(x, y, xi, yi, 16, 16).fuse(x, y, t).parallel(t);
     g.compute_at(h, yi);
-    wrapper.compute_at(h, yi).tile(_0, _1, xi, yi, 8, 8).fuse(xi, yi, t).vectorize(t, 4);
+    wrapper.compute_at(h, yi).tile(x, y, xi, yi, 8, 8).fuse(xi, yi, t).vectorize(t, 4);
 
     // Check the call graphs.
-    // Expect 'g' to call 'wrapper', 'wrapper' to call 'img_f', 'img_f' to call 'img',
+    // Expect 'g' to call 'wrapper', 'wrapper' to call 'f', 'f' to call nothing,
     // 'h' to call 'wrapper' and 'g'
-    Module m = h.compile_to_module({h.infer_arguments()});
+    Module m = h.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
         {h.name(), {g.name(), wrapper.name()}},
         {g.name(), {wrapper.name()}},
-        {wrapper.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {wrapper.name(), {f.name()}},
+        {f.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -199,24 +184,20 @@ int global_in_in_wrapper_test() {
     return 0;
 }
 
-int update_defined_after_in_in_wrapper_test() {
-    Func source("source"), g("g");
+int update_defined_after_wrapper_test() {
+    Func f("f"), g("g");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
+    f(x, y) = x + y;
+    g(x, y) = f(x, y);
 
-    g(x, y) = img(x, y);
+    Func wrapper = f.in(g);
 
-    Func wrapper = img.in(g);
-
-    // Update of 'g' is defined after img.in(g) is called. g's updates should
-    // still call img's wrapper.
+    // Update of 'g' is defined after f.in(g) is called. g's updates should
+    // still call f's wrapper.
     RDom r(0, 100, 0, 100);
     r.where(r.x < r.y);
-    g(r.x, r.y) += 2*img(r.x, r.y);
+    g(r.x, r.y) += 2*f(r.x, r.y);
 
     Param<bool> param;
 
@@ -224,24 +205,23 @@ int update_defined_after_in_in_wrapper_test() {
     RVar rxo("rxo"), rxi("rxi");
     g.specialize(param).vectorize(x, 8).unroll(x, 2).split(x, x, xi, 4).parallel(x);
     g.update(0).split(r.x, rxo, rxi, 2).unroll(rxi);
-    Func img_f = img;
-    img_f.compute_root();
-    wrapper.compute_root().vectorize(_0, 8).unroll(_0, 2).split(_0, _0, xi, 4).parallel(_0);
+    f.compute_root();
+    wrapper.compute_root().vectorize(x, 8).unroll(x, 2).split(x, x, xi, 4).parallel(x);
 
     {
         param.set(true);
 
         // Check the call graphs.
         // Expect initialization of 'g' to call 'wrapper' and its update to call
-        // 'wrapper' and 'g', wrapper' to call 'img_f', 'img_f' to call 'img'
+        // 'wrapper' and 'g', wrapper' to call 'f', 'f' to call nothing
         Module m = g.compile_to_module({g.infer_arguments()});
         CheckCalls c;
         m.functions().front().body.accept(&c);
 
         CallGraphs expected = {
             {g.name(), {wrapper.name(), g.name()}},
-            {wrapper.name(), {img_f.name()}},
-            {img_f.name(), {img.name()}},
+            {wrapper.name(), {f.name()}},
+            {f.name(), {}},
         };
         if (check_call_graphs(c.calls, expected) != 0) {
             return -1;
@@ -261,15 +241,15 @@ int update_defined_after_in_in_wrapper_test() {
 
         // Check the call graphs.
         // Expect initialization of 'g' to call 'wrapper' and its update to call
-        // 'wrapper' and 'g', wrapper' to call 'img_f', 'img_f' to call 'img'
+        // 'wrapper' and 'g', wrapper' to call 'f', 'f' to call nothing
         Module m = g.compile_to_module({g.infer_arguments()});
         CheckCalls c;
         m.functions().front().body.accept(&c);
 
         CallGraphs expected = {
             {g.name(), {wrapper.name(), g.name()}},
-            {wrapper.name(), {img_f.name()}},
-            {img_f.name(), {img.name()}},
+            {wrapper.name(), {f.name()}},
+            {f.name(), {}},
         };
         if (check_call_graphs(c.calls, expected) != 0) {
             return -1;
@@ -287,38 +267,33 @@ int update_defined_after_in_in_wrapper_test() {
     return 0;
 }
 
-int rdom_in_wrapper_test() {
-    Func source("source"), g("g"), result("result");
+int rdom_wrapper_test() {
+    Func f("f"), g("g"), result("result");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
-
+    f(x, y) = x + y;
     g(x, y) = 10;
-    g(x, y) += 2 * img(x, x);
+    g(x, y) += 2 * f(x, x);
     RDom r(0, 200, 0, 200);
-    g(r.x, r.y) += 3 * img(r.y, r.y);
+    g(r.x, r.y) += 3 * f(r.y, r.y);
 
     // Make a global wrapper on 'g', so that we can schedule initialization
     // and the update on the same compute level at the global wrapper
     Func wrapper = g.in().compute_root();
     g.compute_at(wrapper, x);
-    Func img_f = img;
-    img_f.compute_root();
+    f.compute_root();
 
     // Check the call graphs.
     // Expect 'wrapper' to call 'g', initialization of 'g' to call nothing
-    // and its update to call 'img_f' and 'g', 'img_f' to call 'img'
-    Module m = wrapper.compile_to_module({wrapper.infer_arguments()});
+    // and its update to call 'f' and 'g', 'f' to call nothing
+    Module m = wrapper.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {g.name(), {img_f.name(), g.name()}},
+        {g.name(), {f.name(), g.name()}},
         {wrapper.name(), {g.name()}},
-        {img_f.name(), {img.name()}},
+        {f.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -332,37 +307,32 @@ int rdom_in_wrapper_test() {
     return 0;
 }
 
-int global_and_custom_in_in_wrapper_test() {
-    Func source("source"), g("g"), result("result");
+int global_and_custom_wrapper_test() {
+    Func f("f"), g("g"), result("result");
     Var x("x"), y("y");
 
-    source(x) = x;
-    ImageParam img(Int(32), 1, "img");
-    Buffer<int> buf = source.realize(200);
-    img.set(buf);
+    f(x) = x;
+    g(x, y) = f(x);
+    result(x, y) = f(x) + g(x, y);
 
-    g(x, y) = img(x);
-    result(x, y) = img(x) + g(x, y);
-
-    Func img_in_g = img.in(g).compute_at(g, x);
-    Func img_wrapper = img.in().compute_at(result, y);
-    Func img_f = img;
-    img_f.compute_root();
+    Func f_in_g = f.in(g).compute_at(g, x);
+    Func f_wrapper = f.in().compute_at(result, y);
+    f.compute_root();
     g.compute_at(result, y);
 
     // Check the call graphs.
-    // Expect 'result' to call 'g' and 'img_wrapper', 'g' to call 'img_in_g',
-    // 'img_wrapper' to call 'f', img_in_g' to call 'img_f', 'f' to call 'img'
-    Module m = result.compile_to_module({result.infer_arguments()});
+    // Expect 'result' to call 'g' and 'f_wrapper', 'g' to call 'f_in_g',
+    // 'f_wrapper' to call 'f', f_in_g' to call 'f', 'f' to call nothing
+    Module m = result.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {result.name(), {g.name(), img_wrapper.name()}},
-        {g.name(), {img_in_g.name()}},
-        {img_wrapper.name(), {img_f.name()}},
-        {img_in_g.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {result.name(), {g.name(), f_wrapper.name()}},
+        {g.name(), {f_in_g.name()}},
+        {f_wrapper.name(), {f.name()}},
+        {f_in_g.name(), {f.name()}},
+        {f.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -378,32 +348,27 @@ int global_and_custom_in_in_wrapper_test() {
 
 
 int wrapper_depend_on_mutated_func_test() {
-    Func source("sourceo"), f("f"), g("g"), h("h");
+    Func e("e"), f("f"), g("g"), h("h");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
-
-    f(x, y) = img(x, y);
+    e(x, y) = x + y;
+    f(x, y) = e(x, y);
     g(x, y) = f(x, y);
     h(x, y) = g(x, y);
 
     Var xo("xo"), xi("xi");
-    Func img_f = img;
-    img_f.compute_root();
+    e.compute_root();
     f.compute_at(g, y).vectorize(x, 8);
     g.compute_root();
-    Func img_in_f = img.in(f);
+    Func e_in_f = e.in(f);
     Func g_in_h = g.in(h).compute_root();
     g_in_h.compute_at(h, y).vectorize(x, 8);
-    img_in_f.compute_at(f, y).split(_0, xo, xi, 8);
+    e_in_f.compute_at(f, y).split(x, xo, xi, 8);
 
     // Check the call graphs.
     // Expect 'h' to call 'g_in_h', 'g_in_h' to call 'g', 'g' to call 'f',
-    // 'f' to call 'img_in_f', img_in_f' to call 'img_f', 'img_f' to call 'img'
-    Module m = h.compile_to_module({h.infer_arguments()});
+    // 'f' to call 'e_in_f', e_in_f' to call 'e', 'e' to call nothing
+    Module m = h.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
@@ -411,9 +376,9 @@ int wrapper_depend_on_mutated_func_test() {
         {h.name(), {g_in_h.name()}},
         {g_in_h.name(), {g.name()}},
         {g.name(), {f.name()}},
-        {f.name(), {img_in_f.name()}},
-        {img_in_f.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {f.name(), {e_in_f.name()}},
+        {e_in_f.name(), {e.name()}},
+        {e.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -427,39 +392,37 @@ int wrapper_depend_on_mutated_func_test() {
     return 0;
 }
 
-int wrapper_on_in_wrapper_test() {
-    Func source("source"), g("g"), h("h");
+int wrapper_on_wrapper_test() {
+    Func e("e"), f("f"), g("g"), h("h");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
+    e(x, y) = x + y;
+    f(x, y) = e(x, y);
+    g(x, y) = f(x, y) + e(x, y);
+    Func f_in_g = f.in(g).compute_root();
+    Func f_in_f_in_g = f.in(f_in_g).compute_root();
+    h(x, y) = g(x, y) + f(x, y) + f_in_f_in_g(x, y);
 
-    g(x, y) = img(x, y) + img(x, y);
-    Func img_in_g = img.in(g).compute_root();
-    Func img_in_img_in_g = img.in(img_in_g).compute_root();
-    h(x, y) = g(x, y) + img(x, y) + img_in_img_in_g(x, y);
-
-    Func img_f = img;
-    img_f.compute_root();
+    e.compute_root();
+    f.compute_root();
     g.compute_root();
-    Func img_in_h = img.in(h).compute_root();
+    Func f_in_h = f.in(h).compute_root();
     Func g_in_h = g.in(h).compute_root();
 
     // Check the call graphs.
-    Module m = h.compile_to_module({h.infer_arguments()});
+    Module m = h.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {h.name(), {img_in_h.name(), g_in_h.name(), img_in_img_in_g.name()}},
-        {img_in_h.name(), {img_f.name()}},
+        {h.name(), {f_in_h.name(), g_in_h.name(), f_in_f_in_g.name()}},
+        {f_in_h.name(), {f.name()}},
         {g_in_h.name(), {g.name()}},
-        {g.name(), {img_in_g.name()}},
-        {img_in_g.name(), {img_in_img_in_g.name()}},
-        {img_in_img_in_g.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {g.name(), {e.name(), f_in_g.name()}},
+        {f_in_g.name(), {f_in_f_in_g.name()}},
+        {f_in_f_in_g.name(), {f.name()}},
+        {f.name(), {e.name()}},
+        {e.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -474,39 +437,33 @@ int wrapper_on_in_wrapper_test() {
 }
 
 int wrapper_on_rdom_predicate_test() {
-    Func source("source"), g("g"), h("h");
+    Func f("f"), g("g"), h("h");
     Var x("x"), y("y");
 
-    source(x, y) = x + y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(200, 200);
-    img.set(buf);
-
+    f(x, y) = x + y;
     g(x, y) = 10;
     h(x, y) = 5;
 
     RDom r(0, 100, 0, 100);
-    r.where(img(r.x, r.y) + h(r.x, r.y) < 50);
+    r.where(f(r.x, r.y) + h(r.x, r.y) < 50);
     g(r.x, r.y) += h(r.x, r.y);
 
     Func h_wrapper = h.in().store_root().compute_at(g, r.y);
-    Func img_in_g = img.in(g).compute_at(g, r.x);
-    Func img_f = img;
-    img_f.compute_root();
+    Func f_in_g = f.in(g).compute_at(g, r.x);
+    f.compute_root();
     h.compute_root();
 
     // Check the call graphs.
-    // Expect 'g' to call nothing, update of 'g' to call 'g', img_in_g', and 'h_wrapper',
-    // 'img_in_g' to call 'img_f', 'img_f' to call 'img', 'h_wrapper' to call 'h',
-    // 'h' to call nothing
-    Module m = g.compile_to_module({g.infer_arguments()});
+    // Expect 'g' to call nothing, update of 'g' to call 'g', f_in_g', and 'h_wrapper',
+    // 'f_in_g' to call 'f', 'f' to call nothing, 'h_wrapper' to call 'h', 'h' to call nothing
+    Module m = g.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {g.name(), {g.name(), img_in_g.name(), h_wrapper.name()}},
-        {img_in_g.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {g.name(), {g.name(), f_in_g.name(), h_wrapper.name()}},
+        {f_in_g.name(), {f.name()}},
+        {f.name(), {}},
         {h_wrapper.name(), {h.name()}},
         {h.name(), {}},
     };
@@ -524,36 +481,31 @@ int wrapper_on_rdom_predicate_test() {
     return 0;
 }
 
-int two_fold_in_wrapper_test() {
-    Func source("source"), img_in_output_in_output, img_in_output, output("output");
+int two_fold_wrapper_test() {
+    Func input("input"), input_in_output_in_output, input_in_output, output("output");
     Var x("x"), y("y");
 
-    source(x, y) = 2*x + 3*y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(1024, 1024);
-    img.set(buf);
+    input(x, y) = 2*x + 3*y;
+    input.compute_root();
 
-    Func img_f = img;
-    img_f.compute_root();
-
-    output(x, y) = img(y, x);
+    output(x, y) = input(y, x);
 
     Var xi("xi"), yi("yi");
     output.tile(x, y, xi, yi, 8, 8);
 
-    img_in_output = img.in(output).compute_at(output, x).vectorize(_0).unroll(_1);
-    img_in_output_in_output = img_in_output.in(output).compute_at(output, x).unroll(_0).unroll(_1);
+    input_in_output = input.in(output).compute_at(output, x).vectorize(x).unroll(y);
+    input_in_output_in_output = input_in_output.in(output).compute_at(output, x).unroll(x).unroll(y);
 
     // Check the call graphs.
-    Module m = output.compile_to_module({output.infer_arguments()});
+    Module m = output.compile_to_module({});
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {output.name(), {img_in_output_in_output.name()}},
-        {img_in_output_in_output.name(), {img_in_output.name()}},
-        {img_in_output.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
+        {output.name(), {input_in_output_in_output.name()}},
+        {input_in_output_in_output.name(), {input_in_output.name()}},
+        {input_in_output.name(), {input.name()}},
+        {input.name(), {}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -567,44 +519,39 @@ int two_fold_in_wrapper_test() {
     return 0;
 }
 
-int multi_folds_in_wrapper_test() {
-    Func source("source"), img_in_g_in_g, img_in_g, img_in_g_in_g_in_h, img_in_g_in_g_in_h_in_h, g("g"), h("h");
+int multi_folds_wrapper_test() {
+    Func f("f"), f_in_g_in_g, f_in_g, f_in_g_in_g_in_h, f_in_g_in_g_in_h_in_h, g("g"), h("h");
     Var x("x"), y("y");
 
-    source(x, y) = 2*x + 3*y;
-    ImageParam img(Int(32), 2, "img");
-    Buffer<int> buf = source.realize(1024, 1024);
-    img.set(buf);
+    f(x, y) = 2*x + 3*y;
+    f.compute_root();
 
-    Func img_f = img;
-    img_f.compute_root();
-
-    g(x, y) = img(y, x);
+    g(x, y) = f(y, x);
 
     Var xi("xi"), yi("yi");
-    g.compute_root().tile(x, y, xi, yi, 8, 8);
+    g.compute_root().tile(x, y, xi, yi, 8, 8).vectorize(xi).unroll(yi);
 
-    img_in_g = img.in(g).compute_root().tile(_0, _1, xi, yi, 8, 8).vectorize(xi).unroll(yi);
-    img_in_g_in_g = img_in_g.in(g).compute_root().tile(_0, _1, xi, yi, 8, 8).unroll(xi).unroll(yi);
+    f_in_g = f.in(g).compute_root().tile(x, y, xi, yi, 8, 8).vectorize(xi).unroll(yi);
+    f_in_g_in_g = f_in_g.in(g).compute_root().tile(x, y, xi, yi, 8, 8).unroll(xi).unroll(yi);
 
-    h(x, y) = img_in_g_in_g(y, x);
-    img_in_g_in_g_in_h = img_in_g_in_g.in(h).compute_at(h, x).vectorize(_0).unroll(_1);
-    img_in_g_in_g_in_h_in_h = img_in_g_in_g_in_h.in(h).compute_at(h, x).unroll(_0).unroll(_1);
+    h(x, y) = f_in_g_in_g(y, x);
+    f_in_g_in_g_in_h = f_in_g_in_g.in(h).compute_at(h, x).vectorize(x).unroll(y);
+    f_in_g_in_g_in_h_in_h = f_in_g_in_g_in_h.in(h).compute_at(h, x).unroll(x).unroll(y);
     h.compute_root().tile(x, y, xi, yi, 8, 8);
 
     Pipeline p({g, h});
-    Module m = p.compile_to_module({p.infer_arguments()}, "");
+    Module m = p.compile_to_module({}, "");
     CheckCalls c;
     m.functions().front().body.accept(&c);
 
     CallGraphs expected = {
-        {g.name(), {img_in_g_in_g.name()}},
-        {img_in_g_in_g.name(), {img_in_g.name()}},
-        {img_in_g.name(), {img_f.name()}},
-        {img_f.name(), {img.name()}},
-        {h.name(), {img_in_g_in_g_in_h_in_h.name()}},
-        {img_in_g_in_g_in_h_in_h.name(), {img_in_g_in_g_in_h.name()}},
-        {img_in_g_in_g_in_h.name(), {img_in_g_in_g.name()}},
+        {g.name(), {f_in_g_in_g.name()}},
+        {f_in_g_in_g.name(), {f_in_g.name()}},
+        {f_in_g.name(), {f.name()}},
+        {f.name(), {}},
+        {h.name(), {f_in_g_in_g_in_h_in_h.name()}},
+        {f_in_g_in_g_in_h_in_h.name(), {f_in_g_in_g_in_h.name()}},
+        {f_in_g_in_g_in_h.name(), {f_in_g_in_g.name()}},
     };
     if (check_call_graphs(c.calls, expected) != 0) {
         return -1;
@@ -625,37 +572,37 @@ int multi_folds_in_wrapper_test() {
 
 int main(int argc, char **argv) {
     printf("Running calling wrap no op test\n");
-    if (calling_in_in_wrapper_no_op_test() != 0) {
+    if (calling_wrapper_no_op_test() != 0) {
         return -1;
     }
 
     printf("Running func wrap test\n");
-    if (func_in_in_wrapper_test() != 0) {
+    if (func_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running multiple funcs sharing wrapper test\n");
-    if (multiple_funcs_sharing_in_wrapper_test() != 0) {
+    if (multiple_funcs_sharing_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running global wrap test\n");
-    if (global_in_in_wrapper_test() != 0) {
+    if (global_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running update is defined after wrap test\n");
-    if (update_defined_after_in_in_wrapper_test() != 0) {
+    if (update_defined_after_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running rdom wrapper test\n");
-    if (rdom_in_wrapper_test() != 0) {
+    if (rdom_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running global + custom wrapper test\n");
-    if (global_and_custom_in_in_wrapper_test() != 0) {
+    if (global_and_custom_wrapper_test() != 0) {
         return -1;
     }
 
@@ -665,7 +612,7 @@ int main(int argc, char **argv) {
     }
 
     printf("Running wrapper on wrapper test\n");
-    if (wrapper_on_in_wrapper_test() != 0) {
+    if (wrapper_on_wrapper_test() != 0) {
         return -1;
     }
 
@@ -675,12 +622,12 @@ int main(int argc, char **argv) {
     }
 
     printf("Running two fold wrapper test\n");
-    if (two_fold_in_wrapper_test() != 0) {
+    if (two_fold_wrapper_test() != 0) {
         return -1;
     }
 
     printf("Running multi folds wrapper test\n");
-    if (multi_folds_in_wrapper_test() != 0) {
+    if (multi_folds_wrapper_test() != 0) {
         return -1;
     }
